requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
flake8==4.0.1
pytest==7.1.2
pytest-cov==2.8.1
autopep8==1.6.0
fastapi==0.78.0
nest_asyncio